@@ -16,70 +19,64 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "~7.2.0",
-    "@angular/cdk": "^7.3.3",
+    "@angular/cdk": "^7.3.7",
     "@angular/common": "~7.2.0",
     "@angular/compiler": "~7.2.0",
     "@angular/core": "~7.2.0",
     "@angular/forms": "~7.2.0",
-    "@angular/http": "~7.2.0",
-    "@angular/material": "^7.3.3",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^7.3.7",
     "@angular/platform-browser": "~7.2.0",
     "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/platform-server": "~7.2.0",
+    "@angular/platform-server": "^7.2.15",
     "@angular/router": "~7.2.0",
-    "@ng-select/ng-select": "^1.4.1",
-    "@types/lodash": "^4.14.110",
-    "ajv": "6.4.0",
-    "angular-datatables": "^6.0.1",
-    "arrive": "2.4.1",
-    "bootstrap": "4.1.0",
-    "bootstrap-material-design": "4.1.1",
-    "bootstrap-notify": "3.1.3",
-    "chartist": "0.11.0",
-    "classlist.js": "1.1.20150312",
-    "core-js": "2.4.1",
-    "echarts": "^4.1.0",
-    "hammerjs": "2.0.8",
-    "jquery": "^3.2.1",
+    "bootstrap": "^4.3.1",
+    "bootstrap-material-design": "^4.1.2",
+    "bootstrap-notify": "^3.1.3",
+    "chartist": "^0.11.2",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^2.5.4",
+    "dotenv": "^8.0.0",
+    "echarts": "^4.2.1",
+    "hammerjs": "^2.0.8",
+    "jquery": "^3.4.1",
     "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.10",
-    "moment": "^2.19.1",
-    "ngx-toastr": "^9.1.2",
-    "perfect-scrollbar": "1.1.0",
-    "popper.js": "1.14.3",
-    "primeicons": "^1.0.0",
-    "primeng": "^7.1.0",
+    "moment": "^2.24.0",
+    "ngx-toastr": "^10.0.4",
+    "node-sass": "^4.12.0",
+    "perfect-scrollbar": "^1.4.0",
+    "popper.js": "^1.15.0",
     "rxjs": "~6.3.3",
     "rxjs-compat": "~6.3.3",
+    "sass-loader": "^7.1.0",
     "tslib": "^1.9.0",
-    "web-animations-js": "2.3.1",
-    "zone.js": "0.8.4"
+    "web-animations-js": "^2.3.1",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.3",
-    "@angular/compiler-cli": "^7.2.6",
-    "@angular/language-service": "5.2.9",
-    "@types/bootstrap": "3.3.32",
-    "@types/chartist": "0.9.34",
-    "@types/datatables.net": "^1.10.14",
-    "@types/echarts": "0.0.13",
-    "@types/googlemaps": "^3.30.8",
-    "@types/jasmine": "2.5.38",
+    "@angular/cli": "~7.3.3",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/bootstrap": "^4.3.0",
+    "@types/chartist": "^0.9.46",
+    "@types/datatables.net": "^1.10.17",
+    "@types/echarts": "^4.1.9",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.29",
-    "@types/node": "6.0.73",
-    "codelyzer": "4.2.1",
-    "dotenv": "^6.2.0",
-    "jasmine-core": "3.1.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "^4.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "protractor": "^5.4.2",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "3.1.6"
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
   }
 }
