@@ -1,26 +1,28 @@
-import { Injectable } from '@angular/core';
-import { CanActivate, CanActivateChild, ActivatedRouteSnapshot, RouterStateSnapshot, Router, ActivatedRoute } from '@angular/router';
+import {Injectable} from '@angular/core';
+import {CanActivate, CanActivateChild, ActivatedRouteSnapshot, RouterStateSnapshot, Router, ActivatedRoute} from '@angular/router';
 
-import { Observable } from 'rxjs/Observable';
+import {Observable} from 'rxjs/Observable';
 
-import { AuthService } from '../auth/services/auth.service';
-import { VerifyScopeService } from '../services/verify-scope.service';
+import {AuthService} from '../auth/services/auth.service';
+import {VerifyScopeService} from '../services/verify-scope.service';
 
 @Injectable()
 export class ScopeGuard implements CanActivate, CanActivateChild {
-    
+
     constructor(
         private auth: AuthService,
         private router: Router,
         private activeroute: ActivatedRoute,
-        private verifyScopesService: VerifyScopeService) { }
+        private verifyScopesService: VerifyScopeService) {
+    }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+
         const userScopes: Array<String> = this.auth.getScopeUser().split(' ');
         if (route.data && route.data.scope) {
-            const expectedScopes = route.data.scope.split(' ');            
-            const permission = this.verifyScopes(expectedScopes,userScopes);
-            if(!permission){
+            const expectedScopes = route.data.scope.split(' ');
+            const permission = this.verifyScopes(expectedScopes, userScopes);
+            if (!permission) {
                 this.router.navigate(['acess-denied']);
             }
             return permission;
@@ -33,7 +35,7 @@ export class ScopeGuard implements CanActivate, CanActivateChild {
         return this.canActivate(route, state);
     }
 
-    verifyScopes(expectedScopes: Array<String>, userScopes: Array<String>): boolean{
-        return this.verifyScopesService.verifyScopes(expectedScopes,userScopes);
+    verifyScopes(expectedScopes: Array<String>, userScopes: Array<String>): boolean {
+        return this.verifyScopesService.verifyScopes(expectedScopes, userScopes);
     }
 }
