@@ -1,6 +1,6 @@
 const fs = require('fs');
 const dotenv = require('dotenv');
-dotenv.load();
+dotenv.config();
 
 const envsPath = "./src/environments";
 
