@@ -1,22 +1,29 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { LoginComponent } from 'app/security/auth/login/login.component';
-import { ChangePasswordComponent } from 'app/security/auth/change-password/change-password.component';
-import { AcessDeniedComponent } from 'app/core/template/acess-denied/acess-denied.component';
-import { NotfoundComponent } from 'app/core/template/page-not-found/page-not-found.component';
+import { AuthGuard } from 'app/security/guards/auth.guard';
+import { ScopeGuard } from 'app/security/guards/scope.guard';
+import { TemplateComponent } from '../template/template-component/template.component';
+import { AcessDeniedComponent } from '../template/acess-denied/acess-denied.component';
+import { NotfoundComponent } from '../template/page-not-found/page-not-found.component';
 
 const routes = [
-  { path: '', redirectTo: '', pathMatch: 'full' },
-  { path: 'auth/login', component: LoginComponent },
-  { path: 'auth/change', component: ChangePasswordComponent },
+  { path: 'auth', loadChildren: 'app/security/auth/auth.module#AuthModule' },
+  {
+    path: '',
+    component: TemplateComponent, canActivate: [AuthGuard, ScopeGuard], canActivateChild: [AuthGuard, ScopeGuard],
+    children: [
+      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
+      { path: 'dashboard', loadChildren: 'app/modules/dashboard/dashboard.module#DashboardModule' },
+      { path: 'ui', loadChildren: 'app/modules/admin/admin.module#AdminModule' },
+      { path: 'patients', loadChildren: 'app/modules/patient/patient.module#PatientModule' },
+      { path: 'pilotstudies', loadChildren: 'app/modules/pilot-study/pilot-study.module#PilotStudyModule' },
+      { path: 'evaluations', loadChildren: 'app/modules/evaluation/evaluation.module#EvaluationModule' }
+    ]
+  },
   { path: 'acess-denied', component: AcessDeniedComponent },
   { path: 'page-not-found', component: NotfoundComponent },
-  { path: '**', redirectTo: 'page-not-found' },
-  // {
-  //   path: 'admin',
-  //   loadChildren: 'app/modules/admin/admin.modules#AdminRoutingModule'
-  // }
+  { path: '**', redirectTo: 'page-not-found' }
 ];
 @NgModule({
   imports: [
