@@ -4,42 +4,46 @@ import { RouterModule } from '@angular/router';
 import { AdministratorsComponent } from '../administrators/administrators.component';
 import { HealthProfessionalComponent } from '../health-professionals/health-professionals.component';
 import { MyprofileComponent } from '../myprofile/myprofile.component';
-import { ScopeGuard } from 'app/security/guards/scope.guard';
-import { AuthGuard } from 'app/security/guards/auth.guard';
-import { DashboardComponent } from 'app/modules/dashboard/dashboard-component/dashboard.component';
-import { TemplateComponent } from 'app/core/template/template-component/template.component';
 import { MypilotstudiesComponent } from '../mypilotstudies/mypilotstudies.component';
+import { EditMypilotComponent } from '../edit-mypilot/edit-mypilot.component';
+import {MyevaluationsComponent} from "../myevaluations/myevaluations.component";
 
 const routes = [
+  { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
   {
-    path: '',
-    component: TemplateComponent, canActivate: [AuthGuard, ScopeGuard], canActivateChild: [AuthGuard, ScopeGuard],
-    children: [
-      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
-      { path: 'dashboard', component: DashboardComponent, data: {} },
-      {
-        path: 'myprofile',
-        component: MyprofileComponent,
-        data: { scope: "admin:readAll admin:updateAll healthprofessional:read healthprofessional:update" }
-      },
-      {
-        path: 'administrators',
-        component: AdministratorsComponent,
-        data: { scope: "admin:create admin:deleteAll admin:readAll admin:updateAll" }
-      },
-      {
-        path: 'healthprofessionals',
-        component: HealthProfessionalComponent,
-        data: { scope: "healthprofessional:create healthprofessional:deleteAll healthprofessional:readAll healthprofessional:updateAll" }
-      },
-      {
-        path: 'mystudies/:userId',
-        component: MypilotstudiesComponent,
-        data: { scope: "" }
-      },
-      { path: '**', redirectTo: 'page-not-found' },
-    ]
-  }
+    path: 'myprofile',
+    component: MyprofileComponent,
+    data: { scope: "" }
+  },
+  {
+    path: 'administrators',
+    component: AdministratorsComponent,
+    data: { scope: "admins:create admins:delete admins:readAll admins:update" }
+  },
+  {
+    path: 'healthprofessionals',
+    component: HealthProfessionalComponent,
+    data: {
+      scope: "healthprofessionals:create healthprofessionals:readAll"
+    }
+  },
+  {
+    path: 'mystudies',
+    component: MypilotstudiesComponent,
+    data: { scope: "" }
+  },
+  {
+    path: 'mystudies/:userId/:pilotstudy_id',
+    component: EditMypilotComponent,
+    data: { scope: "" }
+  },
+  {
+    path: 'myevaluations',
+    component: MyevaluationsComponent,
+    data: { scope: "" }
+  },
+  { path: '**', redirectTo: 'page-not-found' },
+
 ];
 
 @NgModule({
