@@ -1,12 +1,17 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
+
+
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatSelectModule } from '@angular/material/select';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+
 import { TemplateModule } from 'app/core/template/template.module';
 import { AdminRoutingModule } from './admin-routing/admin-routing.module';
-
 import { AdministratorsComponent } from './administrators/administrators.component';
 import { HealthProfessionalComponent } from './health-professionals/health-professionals.component';
 import { ModalUserComponent } from './modal-user/modal-user.component';
@@ -15,38 +20,51 @@ import { AdminService } from './services/admin.service';
 import { HealthProfessionalService } from './services/health-professional.service';
 import { UserService } from './services/users.service';
 import { SharedModule } from 'app/shared/shared.module';
-import { PilotStudyModule } from '../pilot-study/pilot-study.module';
 import { MypilotstudiesComponent } from './mypilotstudies/mypilotstudies.component';
-import { MatPaginatorModule } from '@angular/material/paginator';
+import { EditMypilotComponent } from './edit-mypilot/edit-mypilot.component';
+import {MatNativeDateModule, MatTableModule} from '@angular/material';
+import { HaniotTableComponent } from './haniot-table/haniot-table.component';
+import { MyevaluationsComponent } from './myevaluations/myevaluations.component';
+import {EvaluationModule} from "../evaluation/evaluation.module";
 
 @NgModule({
   imports: [
-    BrowserModule,
+    CommonModule,
     FormsModule,
     ReactiveFormsModule,
     RouterModule,
-    CommonModule,    
-    
+
     AdminRoutingModule,
     TemplateModule,
     SharedModule,
-    PilotStudyModule,
+    EvaluationModule,
 
+    MatPaginatorModule,
+    MatSelectModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatFormFieldModule,
     MatSlideToggleModule,
-    MatPaginatorModule
+    MatTableModule,
   ],
   declarations: [
-    AdministratorsComponent,   
+    AdministratorsComponent,
     HealthProfessionalComponent,
     ModalUserComponent,
     MyprofileComponent,
-    MypilotstudiesComponent    
+    MypilotstudiesComponent,
+    EditMypilotComponent,
+    HaniotTableComponent,
+    MyevaluationsComponent
 
-  ], 
-  providers:[
+  ],
+  providers: [
     UserService,
     AdminService,
     HealthProfessionalService
+  ],
+  exports: [
+    MypilotstudiesComponent
   ]
 })
 export class AdminModule { }
