@@ -1,18 +1,19 @@
-import { Component } from '@angular/core';
+import { Component, AfterViewInit, OnInit, AfterViewChecked } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
 
 import { ToastrService } from 'ngx-toastr';
 
 import { AdminService } from '../services/admin.service';
 import { ModalService } from 'app/shared/shared-components/haniot-modal/service/modal.service';
-import { IUser, Admin } from '../models/users.models';
+import { IUser, Admin } from '../models/users';
+import { LoadingService } from 'app/shared/shared-components/loading-component/service/loading.service';
 
 @Component({
   selector: 'app-administrators',
   templateUrl: './administrators.component.html',
   styleUrls: ['./administrators.component.scss']
 })
-export class AdministratorsComponent {
+export class AdministratorsComponent implements AfterViewChecked {
   userEdit: IUser = new Admin();
   admins: Array<IUser> = [];
   errorCredentials = false;
@@ -22,21 +23,24 @@ export class AdministratorsComponent {
   limit: number = 5;
   length: number;
 
+  search: string;
+
   constructor(
     private adminService: AdminService,
     private toastr: ToastrService,
-    private modalService: ModalService) {
+    private modalService: ModalService,
+    private loadinService: LoadingService) {
     this.getAllAdministrators();
     /* Verificando a quantidade total de cuidadores cadastrados */
     this.calcLengthAdministrators();
   }
 
-
   getAllAdministrators() {
-    this.adminService.getAll(this.page, this.limit)
+    this.adminService.getAll(this.page, this.limit, this.search)
       .then(admins => {
         this.admins = admins;
         this.calcLengthAdministrators();
+        this.loadinService.close();
       })
       .catch((errorResponse: HttpErrorResponse) => {
         if (errorResponse.status === 401) {
@@ -56,15 +60,15 @@ export class AdministratorsComponent {
         this.modalService.close('modalUser');
         this.modalService.close('modalLoading');
       })
-      .catch((errorResponse: HttpErrorResponse) => {        
+      .catch((errorResponse: HttpErrorResponse) => {
         if (errorResponse.status == 409 &&
           errorResponse.error.code == 409 &&
           errorResponse.error.message == 'A registration with the same unique data already exists!') {
           this.toastr.error('Email já cadastrado');
-        }else{
+        } else {
           this.toastr.error('Não foi possível criar administrador!');
         }
-        this.modalService.actionNotExecuted('modalUser',event,errorResponse.error);
+        this.modalService.actionNotExecuted('modalUser', event, errorResponse.error);
         if (errorResponse.status === 401) {
           this.errorCredentials = true;
         }
@@ -84,10 +88,10 @@ export class AdministratorsComponent {
           errorResponse.error.code == 409 &&
           errorResponse.error.message == 'A registration with the same unique data already exists!') {
           this.toastr.error('Email já cadastrado');
-        }else{
+        } else {
           this.toastr.error('Não foi possível atualizar administrador!');
         }
-        this.modalService.actionNotExecuted('modalUserEdit', event,errorResponse.error);        
+        this.modalService.actionNotExecuted('modalUserEdit', event, errorResponse.error);
         if (errorResponse.status === 401) {
           this.errorCredentials = true;
         }
@@ -108,6 +112,7 @@ export class AdministratorsComponent {
   paginationEvent(event) {
     this.page = event.page;
     this.limit = event.limit;
+    this.search = event.search;
     this.getAllAdministrators();
   }
 
@@ -121,4 +126,9 @@ export class AdministratorsComponent {
         //console.log('Error ao buscar administradores!', errorResponse);
       });
   }
+
+  ngAfterViewChecked() {
+    this.loadinService.close();
+  }
+
 }
