@@ -1,115 +1,168 @@
-import { Component, OnInit } from '@angular/core';
-import { PageEvent } from '@angular/material/paginator';
-import { PilotStudy } from 'app/modules/pilot-study/models/pilot.study';
-import { PilotStudyService } from 'app/modules/pilot-study/services/pilot-study.service';
-import { ActivatedRoute } from '@angular/router';
-import { ModalService } from 'app/shared/shared-components/haniot-modal/service/modal.service';
+import {Component, OnInit, AfterViewInit, AfterViewChecked} from '@angular/core';
+import {PageEvent} from '@angular/material/paginator';
+import {Router} from '@angular/router';
+
+import {ToastrService} from 'ngx-toastr';
+
+import {PilotStudyService} from 'app/modules/pilot-study/services/pilot-study.service';
+import {ModalService} from 'app/shared/shared-components/haniot-modal/service/modal.service';
+import {LoadingService} from 'app/shared/shared-components/loading-component/service/loading.service';
+import {PilotStudy} from 'app/modules/pilot-study/models/pilot.study';
+import {MatTableDataSource} from "@angular/material";
 
 @Component({
-  selector: 'mypilotstudies',
-  templateUrl: './mypilotstudies.component.html',
-  styleUrls: ['./mypilotstudies.component.scss']
+    selector: 'mypilotstudies',
+    templateUrl: './mypilotstudies.component.html',
+    styleUrls: ['./mypilotstudies.component.scss']
 })
-export class MypilotstudiesComponent implements OnInit {
-  userId: string;
-  // MatPaginator Inputs
-  pageSizeOptions: number[] = [5, 10, 25, 100];
+export class MypilotstudiesComponent implements OnInit, AfterViewChecked {
 
-  // MatPaginator Output
-  pageEvent: PageEvent;
+    userId: string;
+    // MatPaginator Inputs
+    pageSizeOptions: number[] = [10, 25, 100];
 
-  /* Controles de paginação */
-  page: number = 1;
-  limit: number = 5;
-  length: number;
+    // MatPaginator Output
+    pageEvent: PageEvent;
 
-  list: Array<PilotStudy>;
-  search: string;
-  searchTime;
+    /* Controles de paginação */
+    page: number = 1;
+    limit: number = 10;
+    length: number;
 
-  cacheStudyIdRemove: string;
+    list: Array<PilotStudy>;
+    search: string;
+    searchTime;
 
-  constructor(
-    private pilotStudyService: PilotStudyService,
-    private  modalService: ModalService,
-    private activeRouter: ActivatedRoute
-  ) { }
+    cacheStudyIdRemove: string;
 
-  ngOnInit() {
-    this.activeRouter.paramMap.subscribe((params) => {
-      this.userId = params.get('userId');
-      this.getAllPilotStudies();
-      this.getLengthPilotStudies();
-    });
-    this.getAllPilotStudies();
-    this.getLengthPilotStudies();
-  }
+    listOfStudiesIsEmpty: boolean;
 
-  getAllPilotStudies() {
-    this.userId = atob(localStorage.getItem('user'));
-    this.pilotStudyService.getAllByUserId(this.userId, this.page, this.limit)
-      .then(studies => {
-        this.list = studies;
-      })
-      .catch(error => {
-        console.log('Erro ao buscar pilot-studies: ', error);
-      });
-  }
-
-  searchOnSubmit() {
-    clearInterval(this.searchTime);
-    this.searchTime = setTimeout(() => {
-      this.pilotStudyService.getAll()
-        .then(studies => {
-          this.list = studies.filter((study) => {
-            return study.name.search(this.search) != -1;
-          });
-        })
-        .catch(error => {
-          console.log('Erro ao buscar pilot-studies: ', error);
-        });
-    }, 200);
-  }
-
-  getIndex(index: number): number {
-    if (this.search) {
-      return null;
+    constructor(
+        private pilotStudyService: PilotStudyService,
+        private loadinService: LoadingService,
+        private router: Router,
+        private modalService: ModalService,
+        private toastService: ToastrService
+    ) {
+        this.list = new Array<PilotStudy>();
+        this.listOfStudiesIsEmpty = false;
     }
-    const size = this.pageEvent && this.pageEvent.pageSize ? this.pageEvent.pageSize : this.limit;
 
-    if (this.pageEvent && this.pageEvent.pageIndex) {
-      return index + 1 + size * this.pageEvent.pageIndex;
+    ngOnInit() {
+        this.loadUserId();
+        this.getAllPilotStudies();
+        this.getLengthPilotStudies();
     }
-    else {
-      return index + Math.pow(size, 1 - 1);
-    }
-  }
 
-  clickPagination(event) {
-    this.pageEvent = event;
-    this.page = event.pageIndex + 1;
-    this.limit = event.pageSize;
-    this.getAllPilotStudies();
-  }
-
-  getLengthPilotStudies() {
-    if (this.userId) {
-      this.pilotStudyService.getAllByUserId(this.userId)
-        .then(studies => {
-          this.length = studies.length;
-        })
-        .catch(error => {
-          console.log('Erro ao buscar pilot-studies: ', error);
-        });
-    } else {
-      this.pilotStudyService.getAll()
-        .then(studies => {
-          this.length = studies.length;
-        })
-        .catch(error => {
-          console.log('Erro ao buscar pilot-studies: ', error);
-        });
+    loadUserId() {
+        this.userId = atob(localStorage.getItem('user'));
     }
-  }
+
+    getAllPilotStudies() {
+        this.userId = atob(localStorage.getItem('user'));
+        this.pilotStudyService.getAllByUserId(this.userId, this.page, this.limit, this.search)
+            .then(studies => {
+                this.list = studies;
+                if (studies.length) {
+                    this.listOfStudiesIsEmpty = false;
+                } else {
+                    this.listOfStudiesIsEmpty = true;
+                }
+            })
+            .catch(error => {
+                this.listOfStudiesIsEmpty = true;
+                // console.log('Erro ao buscar pilot-studies: ', error);
+            });
+    }
+
+    searchOnSubmit() {
+        clearInterval(this.searchTime);
+        this.searchTime = setTimeout(() => {
+            this.pilotStudyService.getAllByUserId(this.userId, this.page, this.limit, this.search)
+                .then(studies => {
+                    this.list = studies;
+                    this.getLengthPilotStudies();
+                })
+                .catch(error => {
+                    // console.log('Erro ao buscar pilot-studies: ', error);
+                });
+        }, 200);
+    }
+
+    getIndex(index: number): number {
+        if (this.search) {
+            return null;
+        }
+        const size = this.pageEvent && this.pageEvent.pageSize ? this.pageEvent.pageSize : this.limit;
+
+        if (this.pageEvent && this.pageEvent.pageIndex) {
+            return index + 1 + size * this.pageEvent.pageIndex;
+        } else {
+            return index + Math.pow(size, 1 - 1);
+        }
+    }
+
+    clickPagination(event) {
+        this.pageEvent = event;
+        this.page = event.pageIndex + 1;
+        this.limit = event.pageSize;
+        this.getAllPilotStudies();
+    }
+
+    getLengthPilotStudies() {
+        if (this.userId) {
+            this.pilotStudyService.getAllByUserId(this.userId, undefined, undefined, this.search)
+                .then(studies => {
+                    this.length = studies.length;
+                })
+                .catch(error => {
+                    // console.log('Erro ao buscar pilot-studies: ', error);
+                });
+        } else {
+            this.pilotStudyService.getAll()
+                .then(studies => {
+                    this.length = studies.length;
+                })
+                .catch(error => {
+                    // console.log('Erro ao buscar pilot-studies: ', error);
+                });
+        }
+    }
+
+    // newPilotStudy() {
+    //     this.router.navigate(['/pilotstudies/new']);
+    // }
+
+    removeStudy() {
+        if (this.cacheStudyIdRemove) {
+            this.pilotStudyService.remove(this.cacheStudyIdRemove)
+                .then(() => {
+                    this.toastService.info('Estudo removido com sucesso!');
+                    this.getAllPilotStudies();
+                    this.getLengthPilotStudies();
+                    this.closeModalConfirmation();
+                })
+                .catch(error => {
+                    this.toastService.error('Não foi possível remover estudo!');
+                });
+        } else {
+            this.toastService.error('Não foi possível remover estudo!');
+        }
+    }
+
+    openModalConfirmation(id: string) {
+        this.cacheStudyIdRemove = id;
+        this.modalService.open('modalConfirmation');
+    }
+
+    closeModalConfirmation() {
+        this.cacheStudyIdRemove = '';
+        this.modalService.close('modalConfirmation');
+    }
+
+    ngAfterViewChecked() {
+        this.loadinService.close();
+    }
 
 }
+
