@@ -4,8 +4,9 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { ToastrService } from 'ngx-toastr';
 import { AdminService } from '../services/admin.service';
 import { HealthProfessionalService } from '../services/health-professional.service';
-import { IUser, HealtArea } from '../models/users.models';
+import { IUser, HealtArea } from '../models/users';
 import { UserService } from '../services/users.service';
+import { AuthService } from 'app/security/auth/services/auth.service';
 
 @Component({
   selector: 'app-myprofile',
@@ -31,32 +32,34 @@ export class MyprofileComponent implements OnInit {
     private adminService: AdminService,
     private healthService: HealthProfessionalService,
     private userService: UserService,
+    private authService: AuthService,
     private toastr: ToastrService
   ) { }
 
   ngOnInit() {
+    this.typeUser = this.authService.decodeToken().sub_type;
     this.getUser();
   }
 
   getUser() {
     this.userId = atob(localStorage.getItem('user'));
-    this.typeUser = localStorage.getItem('typeUser');
-    if (this.typeUser == 'Admin') {
-      this.adminService.getById(this.userId)
-        .then(admin => this.user = admin)
-        .catch(HttpError => {
-          console.log('Não foi possível carregar usuário logado!', HttpError);
-        });
-    } else if (this.typeUser == 'HealthProfessional') {
-      this.healthService.getById(this.userId)
-        .then(healthprofessional => this.user = healthprofessional)
-        .catch(HttpError => {
-          console.log('Não foi possível carregar usuário logado!', HttpError);
-        });
-    } else {
-      this.userService.getTypeUserAndSetLocalStorage(this.userId);
-      this.getUser();
+    switch (this.typeUser) {
+      case 'admin':
+        this.adminService.getById(this.userId)
+          .then(admin => this.user = admin)
+          .catch(HttpError => {
+            console.log('Não foi possível carregar usuário logado!', HttpError);
+          });
+        break;
+      case 'health_professional':
+        this.healthService.getById(this.userId)
+          .then(healthprofessional => this.user = healthprofessional)
+          .catch(HttpError => {
+            // console.log('Não foi possível carregar usuário logado!', HttpError);
+          });
+        break;
     }
+
   }
 
   enabledEdit() {
@@ -65,34 +68,38 @@ export class MyprofileComponent implements OnInit {
   }
 
   onSubmit(form) {
-    if (this.typeUser == 'Admin') {
-      const admin = form.value;
-      admin.id = atob(localStorage.getItem('user'));
-      this.adminService.update(admin)
-        .then(() => {
-          this.getUser();
-          this.toastr.info('Informações atualizadas!');
-          this.visibilityButtonSave = false;
-          this.disabledButtonEdit = false;
-        })
-        .catch((errorResponse: HttpErrorResponse) => {
-          this.toastr.error('Não foi possível atualizar informações!');
-        });
-    } else if (this.typeUser == 'HealthProfessional') {
-      const healthProfessional = form.value;
-      healthProfessional.id = atob(localStorage.getItem('user'));
-      this.healthService.update(healthProfessional)
-        .then(() => {
-          this.getUser();
-          this.toastr.info('Informações atualizadas!');
-          this.visibilityButtonSave = false;
-          this.disabledButtonEdit = false;
-        })
-        .catch((errorResponse: HttpErrorResponse) => {
-          this.toastr.error('Não foi possível atualizar informações!');
-        });
+    switch (this.typeUser) {
+      case 'admin':
+        const admin = form.value;
+        admin.id = atob(localStorage.getItem('user'));
+        this.adminService.update(admin)
+          .then(() => {
+            this.getUser();
+            this.toastr.info('Informações atualizadas!');
+            this.visibilityButtonSave = false;
+            this.disabledButtonEdit = false;
+          })
+          .catch((errorResponse: HttpErrorResponse) => {
+            this.toastr.error('Não foi possível atualizar informações!');
+          });
+        break;
+      case 'health_professional':
+        const healthProfessional = form.value;
+        healthProfessional.id = atob(localStorage.getItem('user'));
+        this.healthService.update(healthProfessional)
+          .then(() => {
+            this.getUser();
+            this.toastr.info('Informações atualizadas!');
+            this.visibilityButtonSave = false;
+            this.disabledButtonEdit = false;
+          })
+          .catch((errorResponse: HttpErrorResponse) => {
+            this.toastr.error('Não foi possível atualizar informações!');
+          });
+        break;
     }
   }
+
   onChangePassword(form) {
     this.userService.changePassword(this.userId, form.value)
       .then(() => {
@@ -100,7 +107,7 @@ export class MyprofileComponent implements OnInit {
         form.reset();
       })
       .catch(HttpError => {
-        console.log('Não foi possível mudar a senha!', HttpError);
+        // console.log('Não foi possível mudar a senha!', HttpError);
         this.toastr.error('Não foi posível mudar sua senha!');
         if (HttpError.error.code == 400 && HttpError.error.message == "Password does not match") {
           form.controls['old_password'].setErrors({ 'incorrect': true });
