@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 
 import { environment } from 'environments/environment';
-import { IUser, Admin } from '../models/users.models';
+import { IUser, Admin } from '../models/users';
 
 @Injectable()
 export class AdminService {
@@ -15,14 +15,21 @@ export class AdminService {
       .toPromise();
   }
 
-  getAll(page?: number, limit?: number): Promise<IUser[]> {
+  getAll(page?: number, limit?: number, search?: string): Promise<IUser[]> {
     let myParams = new HttpParams();
 
-    if (page && limit) {
-      myParams = new HttpParams()
-        .set("page", String(page))
-        .set("limit", String(limit))
-        .set("sort", 'created_a');
+    if (page) {
+      myParams = myParams.append("page", String(page));
+    }
+
+    if (limit) {
+      myParams = myParams.append("limit", String(limit));
+    } else {
+      myParams = myParams.append("limit", String(Number.MAX_SAFE_INTEGER));
+    }
+
+    if (search) {
+      myParams = myParams.append("?email", '*' + search + '*');
     }
 
     const url = `${environment.api_url}/users/admins`;
@@ -37,15 +44,19 @@ export class AdminService {
   }
 
   update(administrator: Admin): Promise<boolean> {
-    return this.getById(administrator.id)
-      .then(adminOld => {
-        Object.keys(adminOld).forEach(key => {
-          if (adminOld[key] == administrator[key] && key != 'id') {
-            delete administrator[key];
-          }
-        });
-        return this.http.patch<any>(`${environment.api_url}/users/admins/${administrator.id}`, administrator)
-          .toPromise();
-      });
+
+    return this.http.patch<any>(`${environment.api_url}/users/admins/${administrator.id}`, administrator)
+      .toPromise();
+
+    // return this.getById(administrator.id)
+    //   .then(adminOld => {
+    //     Object.keys(adminOld).forEach(key => {
+    //       if (adminOld[key] == administrator[key] && key != 'id') {
+    //         delete administrator[key];
+    //       }
+    //     });
+    //     return this.http.patch<any>(`${environment.api_url}/users/admins/${administrator.id}`, administrator)
+    //       .toPromise();
+    //   });
   }
 }
