@@ -23,11 +23,6 @@ export class AreaGraphComponent implements OnInit {
     legend: {
       data: ['1', '2', '3', '4', '5']
     },
-    toolbox: {
-      feature: {
-        saveAsImage: {}
-      }
-    },
     grid: {
       left: '3%',
       right: '4%',
@@ -52,28 +47,28 @@ export class AreaGraphComponent implements OnInit {
         type: 'line',
         stack: '总量',
         areaStyle: { normal: {} },
-        data: [120, 132, 101, 134, 90, 230, 210]
+        data: [120]
       },
       {
         name: '联盟广告',
         type: 'line',
         stack: '总量',
         areaStyle: { normal: {} },
-        data: [220, 182, 191, 234, 290, 330, 310]
+        data: [220]
       },
       {
         name: '视频广告',
         type: 'line',
         stack: '总量',
         areaStyle: { normal: {} },
-        data: [150, 232, 201, 154, 190, 330, 410]
+        data: [150]
       },
       {
         name: '直接访问',
         type: 'line',
         stack: '总量',
         areaStyle: { normal: {} },
-        data: [320, 332, 301, 334, 390, 330, 320]
+        data: [320]
       },
       {
         name: '搜索引擎',
@@ -86,7 +81,7 @@ export class AreaGraphComponent implements OnInit {
           }
         },
         areaStyle: { normal: {} },
-        data: [820, 932, 901, 934, 1290, 1330, 1320]
+        data: [820]
       }
     ]
   };
