@@ -1,9 +1,13 @@
-enum ActivityFrequency {
-    one_per_week, two_per_week, three_per_week, four_more_per_week, none
+export enum ActivityFrequency {
+    one_per_week = "one_per_week",
+    two_per_week = "two_per_week",
+    three_per_week = "three_per_week",
+    four_more_per_week = "four_more_per_week",
+    none = "none"
 }
-export class PhysicalActivityHabitsRecord{
+export class PhysicalActivityHabitsRecord {
     id: string;
     created_at: string;
-    school_activity_freq: string;
+    school_activity_freq: ActivityFrequency;
     weekly_activities: Array<string>;
 }
