@@ -10,7 +10,7 @@ export class FeedingRecordService {
   constructor(private http: HttpClient) { }
 
 
-  getById(patientId: string, feedingRecordId: string,): Promise<FeedingHabitsRecord> {
+  getById(patientId: string, feedingRecordId: string, ): Promise<FeedingHabitsRecord> {
     return this.http.get<any>(`${environment.api_url}/patients/${patientId}/feedinghabitsrecords/${feedingRecordId}`)
       .toPromise();
   }
@@ -18,11 +18,14 @@ export class FeedingRecordService {
   getAll(patientId: string, page?: number, limit?: number): Promise<FeedingHabitsRecord[]> {
     let myParams = new HttpParams();
 
-    if (page && limit) {
-      myParams = new HttpParams()
-        .set("page", String(page))
-        .set("limit", String(limit))
-        .set("sort", 'created_a');
+    if (page) {
+      myParams = myParams.append("page", String(page));
+    }
+
+    if (limit) {
+      myParams = myParams.append("limit", String(limit));
+    } else {
+      myParams = myParams.append("limit", String(Number.MAX_SAFE_INTEGER));
     }
 
     const url = `${environment.api_url}/patients/${patientId}/feedinghabitsrecords`;
