@@ -11,7 +11,7 @@ export class MedicalRecordService {
   constructor(private http: HttpClient) { }
 
 
-  getById(patientId: string, medicalRecordId: string,): Promise<MedicalRecord> {
+  getById(patientId: string, medicalRecordId: string, ): Promise<MedicalRecord> {
     return this.http.get<any>(`${environment.api_url}/patients/${patientId}/medicalrecords/${medicalRecordId}`)
       .toPromise();
   }
@@ -19,11 +19,14 @@ export class MedicalRecordService {
   getAll(patientId: string, page?: number, limit?: number): Promise<MedicalRecord[]> {
     let myParams = new HttpParams();
 
-    if (page && limit) {
-      myParams = new HttpParams()
-        .set("page", String(page))
-        .set("limit", String(limit))
-        .set("sort", 'created_a');
+    if (page) {
+      myParams = myParams.append("page", String(page));
+    }
+
+    if (limit) {
+      myParams = myParams.append("limit", String(limit));
+    } else {
+      myParams = myParams.append("limit", String(Number.MAX_SAFE_INTEGER));
     }
 
     const url = `${environment.api_url}/patients/${patientId}/medicalrecords`;
