@@ -2,10 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 
 import { environment } from 'environments/environment';
-import { SleepHabitsRecord } from '../models/sleep';
 import { PhysicalActivityHabitsRecord } from '../models/physicalActivity';
-
-
 
 @Injectable()
 export class PhysicalActivityRecordService {
@@ -13,7 +10,7 @@ export class PhysicalActivityRecordService {
   constructor(private http: HttpClient) { }
 
 
-  getById(patientId: string, physicalActivityRecordId: string,): Promise<PhysicalActivityHabitsRecord> {
+  getById(patientId: string, physicalActivityRecordId: string, ): Promise<PhysicalActivityHabitsRecord> {
     return this.http.get<any>(`${environment.api_url}/patients/${patientId}/physicalactivityhabits/${physicalActivityRecordId}`)
       .toPromise();
   }
@@ -21,11 +18,14 @@ export class PhysicalActivityRecordService {
   getAll(patientId: string, page?: number, limit?: number): Promise<PhysicalActivityHabitsRecord[]> {
     let myParams = new HttpParams();
 
-    if (page && limit) {
-      myParams = new HttpParams()
-        .set("page", String(page))
-        .set("limit", String(limit))
-        .set("sort", 'created_a');
+    if (page) {
+      myParams = myParams.append("page", String(page));
+    }
+
+    if (limit) {
+      myParams = myParams.append("limit", String(limit));
+    } else {
+      myParams = myParams.append("limit", String(Number.MAX_SAFE_INTEGER));
     }
 
     const url = `${environment.api_url}/patients/${patientId}/physicalactivityhabits`;
