@@ -4,15 +4,13 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { environment } from 'environments/environment';
 import { SleepHabitsRecord } from '../models/sleep';
 
-
-
 @Injectable()
 export class SleepRecordService {
 
   constructor(private http: HttpClient) { }
 
 
-  getById(patientId: string, sleepRecordId: string,): Promise<SleepHabitsRecord> {
+  getById(patientId: string, sleepRecordId: string, ): Promise<SleepHabitsRecord> {
     return this.http.get<any>(`${environment.api_url}/patients/${patientId}/sleephabits/${sleepRecordId}`)
       .toPromise();
   }
@@ -20,11 +18,14 @@ export class SleepRecordService {
   getAll(patientId: string, page?: number, limit?: number): Promise<SleepHabitsRecord[]> {
     let myParams = new HttpParams();
 
-    if (page && limit) {
-      myParams = new HttpParams()
-        .set("page", String(page))
-        .set("limit", String(limit))
-        .set("sort", 'created_a');
+    if (page) {
+      myParams = myParams.append("page", String(page));
+    }
+
+    if (limit) {
+      myParams = myParams.append("limit", String(limit));
+    } else {
+      myParams = myParams.append("limit", String(Number.MAX_SAFE_INTEGER));
     }
 
     const url = `${environment.api_url}/patients/${patientId}/sleephabits`;
