@@ -2,14 +2,15 @@ export interface IPatient {
 
     id: string;
     pilotstudy_id: string;
-    first_name: string;
-    last_name: string;
+    name: string;
+    email: string;
     gender: string;
     birth_date: string;
-
+    password?: string;
+    password_confirm?: string;
 }
 
-export enum Gender{
+export enum Gender {
     male = 'male',
     female = 'female'
 }
@@ -18,12 +19,14 @@ export class Patient implements IPatient {
 
     id: string;
     pilotstudy_id: string;
-    first_name: string;
-    last_name: string;
+    name: string;
+    email: string;
     gender: string;
     birth_date: string;
+    password?: string;
+    password_confirm?: string;
 
-    constructor(){}
+    constructor() { }
 
 }
 
