@@ -3,21 +3,23 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { MultiSelectModule } from 'primeng/multiselect';
 import { MatNativeDateModule, MAT_DATE_LOCALE } from '@angular/material/core';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatSelectModule } from '@angular/material/select';
 import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatTabsModule } from '@angular/material';
 
 import { PatientFormComponent } from './patient-form/patient-form.component';
 import { PatientTableComponent } from './patient-table/patient-table.component';
 import { PatientComponentComponent } from './patient-component/patient-component.component';
-import { PatientService } from './services/patient.service';
+
 import { PatientRoutingModule } from './patient-routing/patient-routing.module';
 import { SharedModule } from 'app/shared/shared.module';
 import { ListPilotstudiesComponent } from './list-pilotstudies/list-pilotstudies.component';
 import { ViewHabitsComponent } from './view-habits/view-habits.component';
 import { HabitsModule } from '../habits/habits.module';
+import { MeasurementModule } from '../measurement/measurement.module';
+import { PatientManagerComponent } from './patient-manager/patient-manager.component';
 
 @NgModule({
   declarations: [
@@ -25,25 +27,32 @@ import { HabitsModule } from '../habits/habits.module';
     PatientTableComponent,
     PatientComponentComponent,
     ListPilotstudiesComponent,
-    ViewHabitsComponent
+    ViewHabitsComponent,
+    PatientManagerComponent
   ],
   imports: [
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
     SharedModule,
+
     MatPaginatorModule,
     MatSelectModule,
     MatDatepickerModule,
     MatNativeDateModule,
     MatFormFieldModule,
-    MultiSelectModule,
+    MatTabsModule,
+
     PatientRoutingModule,
-    HabitsModule
+    HabitsModule,
+    MeasurementModule
   ],
   providers: [
-    PatientService,
-    {provide: MAT_DATE_LOCALE, useValue: 'pt-BR'}
+    { provide: MAT_DATE_LOCALE, useValue: 'pt-BR' }
+  ],
+  exports: [
+    PatientTableComponent,
+      ListPilotstudiesComponent
   ]
 })
 export class PatientModule { }
