@@ -1,52 +1,41 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { RouterModule } from '@angular/router';
+
 import { PilotStudyComponentComponent } from 'app/modules/pilot-study/pilot-study-component/pilot-study-component.component';
-import { TemplateComponent } from 'app/core/template/template-component/template.component';
-import { AuthGuard } from 'app/security/guards/auth.guard';
-import { ScopeGuard } from 'app/security/guards/scope.guard';
 import { PilotStudyFormComponent } from '../pilot-study-form/pilot-study-form.component';
-import { ListPilotstudiesComponent } from 'app/modules/patient/list-pilotstudies/list-pilotstudies.component';
+import { PilotStudyViewComponent } from '../pilot-study-view/pilot-study-view.component';
 
-// const routes: Routes = [
-//   {
-//     path: '',
-//     component: PilotStudyComponentComponent
-//   },
-//   { 
-//     path: 'new',
-//     component: PilotStudyComponentComponent,
-//     data: { scope: ""}
-//   },
-//   { 
-//     path: ':pilotStudyId',
-//     component: PilotStudyComponentComponent,
-//     data: { scope: ""}
-//   }
-  
-// ];
+const routes = [
+  {
+    path: '',
+    component: PilotStudyComponentComponent,
+    data: {
+      scope: "pilots:readAll pilots:delete"
+    }
+  },
 
-const routes = [{
-  path: '',
-  component: TemplateComponent, canActivate: [AuthGuard, ScopeGuard], canActivateChild: [AuthGuard, ScopeGuard],
-  children: [
-    {
-      path: 'pilotstudies',
-      component: PilotStudyComponentComponent,
-      data: { scope: "pilotstudy:readAll pilotstudy:delete pilotstudy:deleteAll"}
-    },
-    { 
-      path: 'pilotstudies/new',
-      component: PilotStudyFormComponent,
-      data: { scope: "pilotstudy:create"}
-    },
-    { 
-      path: 'pilotstudies/:pilotStudyId',
-      component: PilotStudyFormComponent,
-      data: { scope: "pilotstudy:update pilotstudy:updateAll"}
-    },
-    { path: '**', redirectTo: 'page-not-found' },
-  ]
-}];
+  {
+    path: 'new',
+    component: PilotStudyFormComponent,
+    data: { scope: "pilots:create" }
+  },
+  {
+    path: ':pilotStudyId',
+    component: PilotStudyFormComponent,
+    data: { scope: "pilots:update" }
+  },
+  {
+    path: ':pilotStudyId/details',
+    component: PilotStudyViewComponent,
+    data: { scope: "pilots:update" }
+  },
+  {
+    path: ':pilotStudyId/:userRequest',
+    component: PilotStudyFormComponent,
+    data: { scope: "pilots:update" }
+  },
+  { path: '**', redirectTo: 'page-not-found' }
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
