@@ -9,14 +9,32 @@ export class VerifyScopeService {
     if (expectedScopes.length === 0) {
       return true;
     }
-    
-    var allowed = expectedScopes.some(function (scope) {
-      return scopes.indexOf(scope) !== -1;
+
+    let result = true;
+
+    expectedScopes.forEach(scope => {
+      if (!scopes.find(user => {
+        return user == scope;
+      })) {
+        result = false;
+      }
     });
 
-    return allowed ?
-      true :
-      false;
+    return result;
   }
+
+  // verifyScopes(expectedScopes: Array<String>, scopes: Array<String>): boolean {
+  //   if (expectedScopes.length === 0) {
+  //     return true;
+  //   }
+
+  //   var allowed = expectedScopes.some(function (scope) {
+  //     return scopes.indexOf(scope) !== -1;
+  //   });
+
+  //   return allowed ?
+  //     true :
+  //     false;
+  // }
 
 }
